shell: change mind about bat theme
And add go guru open defintion in other window for emacs, along with
changing the region highlight color to be clearer for dark themes.

Mulling over this more, it would be nice if there was someway for emacs
to switch over some LIGHT or DARK env for the various color tweaks. As
then inevitably don't transfer to one another.